feat: enabled switchboard link in RWA editor
